backend/pyproject.toml
[project]
name = "scraping-job-ms-backend"
version = "0.2.0"
description = "Google Maps Lead Scraper — Backend Worker API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "sqlmodel>=0.0.16",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "aiohttp>=3.9.0",
    "pandas>=2.1.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests"]
